refactor: remove BatcherType (#97)
Instead of having a `BatcherType`, create the `Batcher`s directly.
This removes one level of indirection.

BREAKING CHANGE: You now construct `Batcher` directly.

The public API allows to construct CPU and GPU based batchers. It is
also possible to set a custom strength, but the default strength
is used by default.

For better backwards compatibility the `pick_gpu()` constructor is
introduced, which picks and arbitrary GPU. This is useful for testing
or if you have a single GPU only. Though the `new()` constructor,
which takes a specific device should be preferred.