Fix Clippy warnings (#98)
* refactor: fix all clippy warnings

Some warnings are not fixed, but silenced. The code wasn't changed
as it adds claroty in its current form.

* refactor: remove some #[allow(clippy::needless_range_loop)]

Switch to itertors where it made sense. In some cases it's not really
possible, or it makes the code way harder to understand.

* chore: upgrade to Rust 1.51 and fix clippy warnings

* refactor: fix all #[allow(clippy::upper_case_acronyms)]

Instead of having Clippy exceptions, rename those structs.