chore: simplify CI
Clippy also runs a full build, hence additional targets to build it are
not needed. Also make sure that the full feature matrix is tested by
clippy.

When running Clippy with `--all-targets`, it includes benchmarks.
`--workspace` makes sure the gbench is included.

With using `resovler = "2"`, we don't need a nightly Rust for building
gbench.