chore: improve Clippy on CI (#92)
* chore: improve Clippy on CI

Clippy is run on CI, but even if there are warnings, the CI doesn't
fail. With this commit the CI fails when there are Clippy warnings.
It also runs Clippy on all the code (also benchmarks and examples)
now.

* fix: fix more clippy warnings

* chore: simplify CI

Clippy also runs a full build, hence additional targets to build it are
not needed. Also make sure that the full feature matrix is tested by
clippy.

When running Clippy with `--all-targets`, it includes benchmarks.
`--workspace` makes sure the gbench is included.

With using `resovler = "2"`, we don't need a nightly Rust for building
gbench.