fix: use correct global work size
In `rust-gpu-tools` v4.0.0 the definition of the global work size was
changed from the OpenCL to the CUDA definition. This means that the
global work size no longer needs to be a multiple of the local work
size, but it is the multiplier of the local work size. So the total
number of threads used is global work size * local work size.