feat: support other fields
It is now possible to use Neptune with Pasta curves on the GPU. In
order to make that possible, two new compile-time features are introduced.
`bls` to use BLS12-381 and `pasta` to use the Pasta curves (Pallas and
Vesta).

BREAKING CHANGE: The API now requires to pass in the field that
should be used. It does no longer default to BLS12-381.

If the GPU is used (enabled by the `cuda` and/or `opencl` features), you
now also need to specify the supported fields, `bls` and/or `pasta`.