feat: use generic field everywhere
Sometimes a concrete implementation from blstrs was used as a field.
This is now only the case for the tests. There is no dependency on
a specific field anymore. This means that neptune can now also be
used with other fields, like the ones from `pasta_curves`.

BREAKING CHANGE: The API now requires to pass in the field that
should be used. It does no longer default to BLS12-381.