fix: update to newer ec-gpu version
The API of ec-gpu changed. The new version makes naming things easier,
as it can directly be derived from the fields. Also helper functions
where introduced so that less boilerplate is needed for generating
a GPU kernel.

A new trait alias called `NeptuneField` is introduced. It changes the
traits, depending on whether GPU support is enabled or not (via the
`cuda`/`opencl` features). This makes the code easier to read and
maintain. Without that trait alias at places like this:

    impl<F, A> Batcher<F, A>
    where
        F: NeptuneField,
        A: Arity<F>,

You would instead need to write:

    impl<
            #[cfg(not(any(feature = "cuda", feature = "opencl")))] F: PrimeField,
            #[cfg(any(feature = "cuda", feature = "opencl"))] F: PrimeField + GpuField,
            A: Arity<F>,
        > Batcher<F, A>

Setting trait bounds with conditional compilation is not supported
in `where` clauses.

This version uses a fork of `pasta_curves`. Once upstream `pasta_curves`
v0.5.0 is released, it will be switched again. This temporary for is
only needed in order to make things work with the `ec-gpu` v0.2. The
fork also contains Serde support.

BREAKING CHANGE: deeper down dependencies changed

While this change itself is not really a breaking change, updating
to the newer `blstrs` version is. The reason is that it depends on
a pinned version of `blst`, which was updated. If anything within
your dependency tree is using a different version of `blstrs` you
will end up hard to resolve dependency issues.

BREAKING CHANGE: environment variables are renamed

`NEPTUNE_GPU_FRAMEWORK` is now called `EC_GPU_FRAMEWORK`,
`NEPTUNE_CUDA_NVCC_ARGS` is now called `EC_GPU_CUDA_NVCC_ARGS`.

BREAKING CHANGE: the `wasm` feature was removed

Due to the upgrade to a newer `bellperson` release, the `wasm` feature
is not longer needed. Compiling for WebAssembly should work out of the
box.